폴더 구조 플랫화
