+++ src/combineCharacter/combineCharacter.spec.ts
