+++ src/combineCharacter/combineCharacter.ts
