+++ src/disassemble/disassemble.spec.ts
