+++ src/disassemble/disassemble.ts
