+++ src/disassembleCompleteCharacter/disassembleCompleteCharacter.spec.ts
