+++ src/disassembleCompleteCharacter/disassembleCompleteCharacter.ts
