+++ src/getChoseong/getChoseong.spec.ts
