+++ src/getChoseong/getChoseong.ts
