+++ src/hasBatchim/hasBatchim.ts
