+++ src/removeLastCharacter/removeLastCharacter.spec.ts
