+++ src/removeLastCharacter/removeLastCharacter.ts
